LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/table_set64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds (table_set64.wast:1) assert_equals: expected false but got true
PASS #3 Reinitialize the default imports
FAIL #4 Test that WebAssembly compilation succeeds (table_set64.wast:1) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 191: Element init_expr must produce an i32 at {loc} expected true got false
FAIL #5 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #6 Test that a WebAssembly code returns a specific result (table_set64.wast:29) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #7 Test that a WebAssembly code returns a specific result (table_set64.wast:30) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:13:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #8 Test that a WebAssembly code returns a specific result (table_set64.wast:31) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:16:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #9 Test that a WebAssembly code returns a specific result (table_set64.wast:32) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #10 Test that a WebAssembly code returns a specific result (table_set64.wast:33) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #11 Test that a WebAssembly code returns a specific result (table_set64.wast:35) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #12 Test that a WebAssembly code returns a specific result (table_set64.wast:36) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #13 Test that a WebAssembly code returns a specific result (table_set64.wast:37) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #14 Test that a WebAssembly code returns a specific result (table_set64.wast:38) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:34:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #15 Test that a WebAssembly code returns a specific result (table_set64.wast:39) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #16 Test that a WebAssembly code traps (table_set64.wast:41) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:40:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #17 Test that a WebAssembly code traps (table_set64.wast:42) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:43:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #18 Test that a WebAssembly code traps (table_set64.wast:43) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:46:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #19 Test that a WebAssembly code traps (table_set64.wast:44) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:49:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #20 Test that a WebAssembly code traps (table_set64.wast:46) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:52:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #21 Test that a WebAssembly code traps (table_set64.wast:47) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:55:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #22 Test that a WebAssembly code traps (table_set64.wast:48) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:58:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
FAIL #23 Test that a WebAssembly code traps (table_set64.wast:49) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_set64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:61:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_set64.wast.js:63:3 expected true got false
